+++ NEWS.md
@@ -19,6 +19,7 @@
 
 **New features**
 
+* [#393](https://github.com/r-geoflow/geoflow/issues/393) Add option to geosapi action to define store base datadir
 * [#417](https://github.com/r-geoflow/geoflow/issues/417) Support surface computation global option
 * [#425](https://github.com/r-geoflow/geoflow/issues/425) Support new Data sourceFid
 * [#426](https://github.com/r-geoflow/geoflow/issues/426) Enable Entity data validator